replace FoldCase by IsText and use simple case folding
The toFoldCase function of Data.Text massively increases compilation times
with recent versions of GHC. In this package the simpler case folding
implementation of the toLower function in Data.Char is sufficient.

This is a backward incompatible API change since the scope of
the respective functions is excludes the ByteString instances of
IsString and FoldCase.